DateBox filter: Improved shortcuts
"+" used to increase the day. There was no "-" to decrease, because
that may be used as date delimiter(?).

I think this change makes the shortcuts more intuitive:
* CursorUp/Down : in/decrease day
* SHIFT+CursorUp/Down : in/decrease month
* PageUp/Down : in/decrease year

In/decreasing the year is only seldomly used, so it's fine to have it
on a less commonly used shortcut. In/decreasing the day/month is handy
much more often and is therefore on the most familiar cursor keys.

Updated help file accordingly.

This change also reduces code duplication.