Keep sidebar at a constant width
Don't resize the sidebar (account list and QuickTracker) when resizing
the window. Keep it at the width of the widest account name, or at least
the width of "Balance: $99,999,999.00" to avoid a linebreak in QuickTracker.
If your networth is 100+ million dollars, check your priorities... :)

Remove no longer needed horizonal scrollbar from the account view.