Manage "Type" in CategoryBox
Instead of having the user fill in the "Type" of transaction
(income (+, type "DEP"), spending (-, type "ATM"), have it be
determined by the chosen category.

In case of transactions without category, make the user choose
the type in the alert asking if there really should be no category
for the transaction.

CatergoryBox now has the public members
* SetType() to set a BString to "DEP" or "ATM"
* GetType() to get the BString ("DEP" or "ATM")

The private member SetTypeFromCategory() sets the type from the
category name of the BTextControl.

Instead of checking on the type in the CheckNumBox (= the "Type"
BTextControl), we now get the type from the CategoryBox (GetType()).
Similarly, we now set the type in the CategoyBox (SetType()),
not the CheckNumBox.

At this time, the "Type" text box is still visible in the GUI.
If all works as planned, it will be removed.