Remove "Type" from GUI
The "Type" (DEP, ATM, XFER) is still kept in the database, managed
by the selected category. This change removes the "Type" textbox
from the GUI, because the user doesn't have to fill it out any more.

In the "Transaction report" is a column showing the type of a transaction.
Instead of showing "ATM, DEP or XFER", show the common terms "Spending,
Income, Transfer".

Adjusted a few column widths of the reports to avoid truncating.