Filter view: Fixed width
The filter view consists of two columns, one for the labels, one
for the menu and text fields.

Set the 2nd column min width to the width of the longest "Period"
menu label + a B_USE_ITEM_SPACING for the menu arrow widget.

Now the sidebar width doesn't jump when choosing a different "Period"
with a longer/shorter label.