Remove code related to "checks"
We've decided in the past that we won't treat "checks" any different
than an income/spending transaction. Several changes in that regard
were already made, e.g. not react on numbers entered into the "Type"
field, which itself was removed.

This commit removes the last remnants of "checks" from the app, including
the last visible place, the Reconcile window.