Add menu to use the selected transaction for a new one
Sometimes you want to enter a transaction that is very similar to
one that already exists.

The new menu "Transaction | Use as new transaction" (ALT+U) copies
the data of the currently selected transaction to fill the fields
of the CheckView.
Only the "Date" is set to today.
The date field gets selected so the user can start editing right away.

When the selection changes, the CheckView fields are cleared.