Allow to open more than one CapitalBe instance
As it may come in handy to compare to ledgers side-by-side.

To not simply open the last opened file from the "lastfile" setting
of the preference file, we save the path that was sent to RefsReceived()
in fFile.

If CapitalBe was launched with a file, we let ReadToRun() handle it.
If we're not launching, i.e. the "Open" menu was invoked, we open a
MainWindow with the file directly in RefsReceived().

ReadToRun() checks if CapitalBe was started with a file, e.g. double-click.
If not, it gets the "lastfile" setting of the preference file, as usual.