Alert when opening an already open ledger
Since CapitalBe can work with several instances/ledgers in parallel,
users should be warned when they try to open an already open ledger.
Changes could be lost, because the last closed ledger window will
overwrite the ledger file.

* Set an attribute "filelock" of the ledger file to "true" when opening
  it, and to "false" when closing.

* When opening a file, check if "filelock" is false. If it isn't,
  pop an alert to warn the user about possible data loss when editing
  the same file in parallel.

As that "filelock" flag could be spuriously "true", e.g. if the app
crashed or was killed before setting the "filelock" attribute to "false,
the BAlert offers to open the file anyway.