Customize window tab's Zoom button
You never need the whole screen width; that only makes all the
textboxes extremely wide. You want to see as many rows of the
transaction list as possible.
So, instead of going full-screen, only maximize the height when
clicking the Zoom button in the window tab.

We save the un-Zoomed size/position of the window and return to
its position and height when clicking Zoom again. We keep the
width of the max-Zoomed state, because you probably want to keep it,
if you changed it while being max-Zoomed.

We save the Zoom state and restore it on the next launch.