Mark future transaction dates as "invalid"
MarkAsInvalid() paints a red border around the BTextControl.
We do that now for dates entered by the user that are in the future.
It's not prohibited - maybe someone's workflow has a use for it,
but we point it out subtly.

Future-dates in the transaction list are using gNegativeColor as
text colour to have them stand out similarly.