Report: Improved summarily display of categories
Print out one row per category, with all the subtotals in columns.

Since we don't know if a category is of type income or expense, we
can end up with an empty row. By adding a row with the initial title
"<emptyrow>", we can later on ignore those when filling the report
grid.

Fixes #123