Fix opening balance
We used to add the "opening balance" without ID, resulting in
having only a single transaction for all accounts. Setting an
opening balance for a new account removed the existing
"opening balance" transaction from the previous account...

Now we add the opening transaction with an ID and only
remove+add the opening balance for non-new accounts (there
may already be an opening balance that got edited).