Fix date-picker confusing American date format
Choosing a date with the calendar button, the selected date was hard-coded
with "." as delimiter as "day.month.year". The later parsing interpreted
this as "month/day/year" when the locale is set to e.g. US American.

This fixes that by creating the date string according to the current locale.
It's not really satisfying - as the whole date-handling back-forth between
BDate and strings of the whole app feels wasteful...
OTOH, adding another locale conversion here isn't that bad, as it's only
invoked once for the date chosen in the calendar window...

Fixes #129